+++ app/[locale]/template/page.tsx
@@ -20,7 +20,7 @@ import { useTemplatesQuery, } from '@/generated/graphql';
 
 // Components
 import PageHeader from '@/components/PageHeader';
-import TemplateListItem from '@/components/TemplateListItem';
+import TemplateSelectListItem from '@/components/TemplateSelectListItem';
 import { ContentContainer, LayoutContainer, } from '@/components/Container';
 import ErrorMessages from '@/components/ErrorMessages';
 
@@ -28,6 +28,7 @@ import ErrorMessages from '@/components/ErrorMessages';
 import { useScrollToTop } from '@/hooks/scrollToTop';
 
 import logECS from '@/utils/clientLogger';
+import { toSentenceCase } from '@/utils/general';
 import { routePath } from '@/utils/routes';
 import {
   TemplateSearchResultInterface,
@@ -112,7 +113,9 @@ const TemplateListPage: React.FC = () => {
       return [item.title,
         contentText,
       item.funder,
-      item.publishStatus
+      item.publishStatus,
+      item.visibility,
+      item.description
       ]
         .filter(Boolean)
         .some(field => field?.toLowerCase().includes(lowerCaseTerm));
@@ -157,7 +160,7 @@ const TemplateListPage: React.FC = () => {
 
 
   useEffect(() => {
-    // Transform templates into format expected by TemplateListItem component
+    // Transform templates into format expected by TemplateSelectListItem component
     if (data && data?.myTemplates) {
       const fetchAllTemplates = async (templates: PaginatedTemplateSearchResultsInterface | null) => {
         const items = templates?.items ?? [];
@@ -165,17 +168,14 @@ const TemplateListPage: React.FC = () => {
           items.map(async (template: TemplateSearchResultInterface | null) => {
             return {
               title: template?.name || "",
-              link: routePath('template.show', { templateId: template?.id ?? '' }),
-              content: template?.description || template?.modified ? (
-                <div>
-                  <p>{template?.description}</p>
-                  <p>Last updated: {(template?.modified) ? formatDate(template?.modified) : null}</p>
-                </div>
-              ) : null, // Set to null if no description or last modified data
+              link: routePath('template.show', { templateId: template?.id ?? '' }) || '',
               funder: template?.ownerDisplayName,
               lastUpdated: (template?.modified) ? formatDate(template?.modified) : null,
+              lastRevisedBy: template?.modifiedByName,
               publishStatus: (template?.isDirty) ? 'Unpublished' : 'Published',
-              defaultExpanded: false
+              publishDate: (template?.latestPublishDate) ? formatDate(template?.latestPublishDate) : null,
+              defaultExpanded: false,
+              visibility: toSentenceCase(template?.visibility ? template?.visibility?.toString() : '')
             }
           }));
 
@@ -299,8 +299,8 @@ const TemplateListPage: React.FC = () => {
                 filteredTemplates.slice(0, visibleCount['filteredTemplates']).map((template, index) => {
                   const isFirstInNextSection = index === visibleCount['filteredTemplates'] - 3;
                   return (
-                    < div ref={isFirstInNextSection ? nextSectionRef : null} key={index} >
-                      <TemplateListItem
+                    <div ref={isFirstInNextSection ? nextSectionRef : null} key={index}>
+                      <TemplateSelectListItem
                         key={index}
                         item={template} />
                     </div>
@@ -323,7 +323,7 @@ const TemplateListPage: React.FC = () => {
                   const isFirstInNextSection = index === visibleCount['templates'] - 3;
                   return (
                     <div ref={isFirstInNextSection ? nextSectionRef : null} key={index}>
-                      <TemplateListItem
+                      <TemplateSelectListItem
                         key={index}
                         item={template} />
                     </div>